mctoexception v1.1.0
This tool converts the contents of a Microsoft Message Compiler (.MC)
file into a set of std::exception classes.  The generated exceptions
accept constructor arguments to represent the insertion strings of the
message and loads the message text using the language of the calling
thread.  License has been switched to the MIT license. Requires .NET
Framework 4.5 or higher.